And bump the minimum Nodejs version to the versions that support this feature.
Dev dependencies are also bumped to the newest ones that can run on said minimal nodejs versions and do not require further changes to the code or tests.

Also added Windows to the CI tests and updated `actions/setup-node` to `v4`. Excluded Node 14 from MacOS, as those binaries are not available anymore.